repalce nav in chatlist, create more folders for styles
