+++ src/components/ConnectLine.js
@@ -3,7 +3,7 @@ import { motion as m } from "framer-motion";
 import { useMemo } from "react";
 import { useSelector } from "react-redux";
 
-import "./../styles/ConnectLine.css";
+import "./../styles/helpers/ConnectLine.css";
 
 export default function ConnectLine() {
   const isSocketConnected = useSelector(getNetworkState);
