+++ src/styles/Main.css
@@ -1,21 +1,8 @@
 main {
-  width: 100vw;
-  height: 91vh;
+  width: 100svw;
+  height: 100svh;
 
   display: flex;
-}
-nav {
-  width: 100%;
-  height: 7vh;
-
-  padding-left: 20px;
-  padding-right: 20px;
-
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-
-  overflow: hidden;
 }
 .nav-logo {
   display: flex;
@@ -29,8 +16,15 @@ nav {
 }
 
 .nav-navigate-bar {
+  padding: 5px 0;
+
   display: flex;
-  justify-content: right;
+  justify-content: space-between;
+  align-items: center;
+}
+.nav-btns {
+  display: flex;
+  justify-content: center;
   align-items: center;
 }
 
@@ -68,8 +62,8 @@ nav {
 }
 
 .nav-navigate-slice {
-  margin-left: calc(var(--font-size) * 0.8125);
-  margin-right: calc(var(--font-size) * 0.8125);
+  margin-left: calc(var(--font-size) * 0.125);
+  margin-right: calc(var(--font-size) * 0.125);
 }
 .nav-navigate-slice span {
   color: var(--text-accent);
